requirements.txt
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.5.5.64
pandas==1.4.2
Pillow==9.2.0
pydicom==2.3.0
streamlit==1.12.0
